refactor: 모바일 네이밍 제거 및 TodoCard 컴포넌트 분리
